pet_projects/zhirkov_low_level_programming_assignments/image_rotation fix: commented code removed; code comments added; try to fix indentation display on github

+++ pet_projects/zhirkov_low_level_programming_assignments/image_rotation/src/image/bmp/bmp_utilities.c
@@ -4,11 +4,10 @@
 
 extern const int_fast8_t PIXEL_SIZE_IN_BYTES;
 
-int_fast8_t bmp_image_width_padding_in_bytes( const uint_fast64_t bmp_image_width )
+uint_fast8_t bmp_image_width_padding_in_bytes( const uint_fast64_t bmp_image_width )
 {
-    return ( 4 - ( bmp_image_width *(uint_fast64_t) PIXEL_SIZE_IN_BYTES ) % 4 ) % 4;
+    return ( 4 - (uint_fast8_t)(( bmp_image_width *(uint_fast64_t) PIXEL_SIZE_IN_BYTES )) % 4 ) % 4;
 }
-
 
 size_t bmp_image_padded_pixel_array_size_in_bytes( const uint32_t bi_width, const uint32_t bi_height )
 {
